pyproject.toml
[tool.poetry]
name = "aws-data-pipeline"
version = "0.1.0"
description = ""
authors = ["ljvrflux <louis@fluxrobotics.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.34.14"
numpy = "^1.26.3"
pyyaml = "^6.0.1"
pandas = "^2.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
